+++ src/lib.rs
@@ -73,7 +73,7 @@
 
  Enable functions, which leverage [`enumset`](https://crates.io/crates/enumset).
  This is especially usefull to get all set status events at once,
- see for example [`serial::Serial::events()`]
+ see for example [`serial::Serial::triggered_events()`]
 
  ### `defmt`